requirements.txt
fastapi
python-dotenv
cohere
chromadb
langchain
python-docx
lxml